requirements.txt
streamlit>=1.30.0
streamlit-folium>=0.18.0
folium>=0.15.0
geopandas>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
shapely>=2.0.0
rasterio>=1.3.0
Pillow>=10.0.0
earthengine-api>=0.1.380
geemap>=0.32.0
langchain>=0.1.0
langchain-core>=0.1.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
fiona>=1.9.0
